Fix Dublin Core (oai_dc) output in OAI-PMH repository...
Note: old versions that worked on MW 1.4 included the page text as
the dc:description field. Then it was broken for a long time. :)
I'm now stripping out the text, as it's not really appropriate
for the description field.

This allows the use of oai_dc format to grab metadata about the pages
without fetching the actual page content, which should be more useful
for those simply wanting a set of page update notifications.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@27892 dd0e9695-b195-4be7-bd10-2dea1a65a6b6