(bug 6199) Generate error for unbalanced <ref>s
This basically uses the patch I posted to that bug two years ago.  It's crude, but it should avoid the most common false positives while hopefully not causing too many false negatives.  It should be possible to refine it to avoid even more false negatives, but for the time being, this will at least prevent most of the constant headaches that newbies get when chunks of articles vanish because they forgot a closing </ref>.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@40998 dd0e9695-b195-4be7-bd10-2dea1a65a6b6