Server side for test run data integration into code review.
Parser tests on trunk can now upload test results to a wiki running CodeReview.
While the tests are running, the run data is marked as in progress and will
display a spinner in the test column on the revision view. After completion,
the result set is uploaded, and success/total counts are displayed, with nice
red digits if there's failures.

Failed tests are listed by name on the revision detail page. Currently details
such as the output diff are not included, but there's a blob field ready to accept such.

The status and result updates are sent to the API, action=codetestupdate. Rather
than forcing the client side to figure out how to log in with an account, I'm
just protecting it with an HMAC which validates that the request is legit; a
shared secret should be configured on the client and server ends.

Probably not too efficient in the list view as it'll pull a query per rev to get
the counts... There's a summary of counts in code_test_run however there may be
multiple rows as this is extended past ParserTests to encompass other test suites
yet to be developed.

Currently no UI for the test suite control data either; and we really need to
clean up how the UI handles branching; we should make interesting branches/directories
easily selectable in the UI so we can follow just trunk, or just some branch, or
whatever.

Note that, for instance, it only really makes sense to run the parser tests on trunk phase3,
not on unrelated stuff in other branches; it'll let you send data for other versions, it
just will be confusing. ;)



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@54066 dd0e9695-b195-4be7-bd10-2dea1a65a6b6