New Extension: FilePageMasking
Allows file description page URLs to be automatically rewritten by Mediawiki so that they don't appear to end in a file extension.  This avoids confusing systems and search engines that use file extensions on URLs to determine behavior.

By default, it converts ".xxx" to "_xxx" for calls to file description pages.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@56925 dd0e9695-b195-4be7-bd10-2dea1a65a6b6