SVGEdit extension: WikiEditor toolbar button integration
Adds a button to the advanced insert section to create a new SVG canvas. After saving, the editor inserts a [[File:foo.svg]] to include it inline.
Currently the filename is generated from a stripped copy of the edited page name, with 'drawing' and the date/time appended, and can't be changed except by renaming the file later.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@83134 dd0e9695-b195-4be7-bd10-2dea1a65a6b6