Based on feedback from folks at enwikinews - having the special page named Special:GoogleNewsSitemap is rather unpopular. Rename to special:NewsFeed.
For now I just made the main en alias for the special page be that. We may
want to rename the extension at some point later.

Also add a version number (Amgine wanted it) Also add my name to the author
list for this extension since I'm feeling a little vain ;)


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@87458 dd0e9695-b195-4be7-bd10-2dea1a65a6b6