[CodeReview] The choice of which SVN adaptor to use is per-installation, not per-repo, therefore it is redundant to report this information for each repository (when being run with the 'all' parameter).  Now we only output the adaptor information the first time that importRepo() is called.
I also added a message so that the repo name is output before anything else.  In some situations it was not being output at all (e.g. if the starting point was invalid) but in general it was being displayed too late in the process, which is very confusing when syncing multiple repos.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@89947 dd0e9695-b195-4be7-bd10-2dea1a65a6b6