* (bug 29564) Fix XML export output when LiquidThreads thread data contains bogus UTF-8
We should still fix the input validation so that signature threads are truncated cleanly etc, but this will fix the XML export of such bogus data by running the bits through UtfNormal::cleanUp(). This applies UTF-8 validation, removes bogus sequences, and also applies NFC transformations, as is already done on the rest of the XML export output.




git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@90723 dd0e9695-b195-4be7-bd10-2dea1a65a6b6