(bug 11685) - Make DynamicPageList limit parser cache to 24 hours by default (configurable) on pages using it's tag.
I think 24 hours is sane, but this should definitly be reviewed by someone familar with performance-y stuff before it gets deployed to make sure it won't blow things up.

(The reason for doing this is DPL does queries of categories, but doesn't check if those categories change, so in extreme cases it could have outdated dpl results on a page for several weeks. In practise this is rarely an issue (on enwikinews anyways), but does cause problems in certain circumstances.)


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@103382 dd0e9695-b195-4be7-bd10-2dea1a65a6b6