(bug 32351) Make #time[l] support explicitly specified timezones. Patch by  Van de Bugger.
This would make something like {{#time:H:i:s| 9:30 January 1, 2012 MST}} convert the time from 9:30 mountatin standard time to whatever it is in UTC.

I made one minor change from the patch on bugzilla in changing an @ to a wfSuppressWarnings. (The @ was already in the code, it wasn't introduced by the patch).


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@105459 dd0e9695-b195-4be7-bd10-2dea1a65a6b6