* (bug 321223) Make horizontal scrolling work on Android browser when there are wide things on the page.
For some reason there was an 'overflow-x: hidden' on the body tag in the global styles in MobileFrontend. This disables horizontal scrolling in the Android browser, making parts of page content that are too wide inaccessible.
Removing this as there seems to be no good reason for it, and have a few requests to make horiz scrolling work on Android from Erik M etc.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@105582 dd0e9695-b195-4be7-bd10-2dea1a65a6b6