Reverting inline commenting from CodeReview
This change is still very buggy and has not recieved any love since initial commit. From the beginning it was an expertimental feature.

Instead of letting this sit in trunk forever, I'm just going to revert it. If you'd like to bring it back, finish it completely or make a branch.

ping r95429, r95435, r95680, r95680, r96173, r99030, r99034, r107068, r107069, r107074


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@108350 dd0e9695-b195-4be7-bd10-2dea1a65a6b6